refactor(business): isolating logic related to summary of payables in a class in business layer
